update task details

+++ src/components/TaskDetailsView/TaskDetailsView.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useCallback, useEffect } from 'react';
+import React, { useState, useCallback, useEffect, useRef } from 'react';
 import {
   DeleteIcon,
   Dropdown,
@@ -21,7 +21,7 @@ interface Props extends FullScreenDialogProps {
   taskLists: Schema$TaskList[];
   currentTaskList: Schema$TaskList | null;
   deleteTask(task: Schema$Task): void;
-  updateTask?(task: Schema$Task): void;
+  updateTask(task: Schema$Task): void;
 }
 
 export function EditTaskButton({ onClick }: { onClick(): void }) {
@@ -46,16 +46,25 @@ export function TaskDetailsView({
 }: Props) {
   const { anchorEl, setAnchorEl, onClose } = useMuiMenu();
   const MenuItem = useMenuItem(handleClose);
+  const notesInputRef = useRef<HTMLTextAreaElement>(null);
+  const shouldDeleteTask = useRef<boolean>(false);
 
   const [task, setTask] = useState(initialTask);
+
   const onExitCallback = useCallback(() => {
-    updateTask && updateTask(task);
+    updateTask(task);
   }, [task, updateTask]);
 
   const deleteTaskCallback = useCallback(() => {
-    deleteTask(task);
+    if (shouldDeleteTask.current) {
+      deleteTask(task);
+    }
+  }, [deleteTask, task]);
+
+  const deleteBtnClickedCallback = useCallback(() => {
+    shouldDeleteTask.current = true;
     handleClose();
-  }, [deleteTask, handleClose, task]);
+  }, [handleClose]);
 
   // FIXME:
   useEffect(() => {
@@ -67,11 +76,12 @@ export function TaskDetailsView({
       className="task-details-view"
       handleClose={handleClose}
       onExit={onExitCallback}
+      onExited={deleteTaskCallback}
       headerComponents={
         <IconButton
           tooltip="Delete"
           icon={DeleteIcon}
-          onClick={deleteTaskCallback}
+          onClick={deleteBtnClickedCallback}
         />
       }
       {...props}
@@ -89,6 +99,8 @@ export function TaskDetailsView({
         className="filled task-details-view-notes-field"
         placeholder="Add details"
         value={task.notes}
+        inputRef={notesInputRef}
+        onClick={() => notesInputRef.current!.focus()}
         onChange={evt => setTask({ ...task, notes: evt.currentTarget.value })}
       />
       <div className="task-details-view-row row-task-list">
